third_party/blink/web_tests/external/wpt/css/css-sizing/contain-intrinsic-size/auto-008-expected.txt

This is a testharness.js-based test.
PASS requestAnimationFrame
FAIL Early ResizeObserver assert_equals: Using last remembered size - clientWidth expected 100 but got 40
PASS Late ResizeObserver
Harness: the test ran to completion.
